formatação dos preços e adicionando o ícone do botão

+++ yahoo-finance-ios/yahoo-finance-ios/Service/YahooModel.swift
@@ -16,8 +16,8 @@ public struct QuoteResponse: Decodable {
 }
 
 public struct Result: Decodable {
-    let bid: Decimal
-    let ask: Decimal
+    let bid: Float
+    let ask: Float
     let currency: String
     let displayName: String
 }
